Refactor ProgressLogger for extensibility
Pulled out an AbstractProgressLoger and refactored project logger to implement that.
This allows other implementations of AbstractProgressLogger that easily substitute a different logging system other than the Picard built in.
